Fixes scoped package management
Not sure if this is the ideal place to fix this, but it is the least intrusive (I think to fix it at its source means that packages would have to be re-published).

The problem is that the packages are asking for scoped packages in a format like this...

    http://localhost:3000/@opendns/grunt-compile-ich/-/@opendns/grunt-compile-ich-1.0.0/hash.tgz

But that was not matching the route pattern.  Anyway, this works for us.